Add support for datetime values as arguments
This patch adds some custom encoding/decoding to the JSON parser to
allow it to work with datetime values.  They will be serialized
according to DATE_PREFIX and DATE_FORMAT in json_parser.py.

This patch also removes the dependency on simplejson in Python 2.6,
although simplejson is still preferred if it's available.