add enqueue & perform timestamps, add job hooks for accessing them
Classes which implement the perform() method can now also implement
before_perform() and after_perform() methods which take a metadata dict.
These contain the args passed in, as well as timestamps for when the job
was enqueued and when it was performed, and whether the job failed and
was retried.