Adding test to check whether pyres identifies a crashed process as a failure (currently failing).
The biggest problem shown by this test is the fact that, even after the child
process has crashed and the worker is ready to process new jobs, the entry
associating the job with the worker remain on Redis.

It also refactors the test for timeout errors on staled workers, making it simpler.