handle jobs queued by Ruby Resque with no module
Ruby compatibility: Resque sends just a class name and not a module name
so if I use Resque to queue a ruby class called "Worker" then Pyres will
throw a "ValueError: Empty module name" exception.  To avoid that, if
there's no module name in the json then we'll use the class name as a
module name.